requirements.txt
environ-config
minio
nltk
python-dotenv
python-twitter
sklearn
